+++ Middlewares/ai-postprocessing-wrapper/app_postprocess_od_st_yolox_uf.c
@@ -26,15 +26,15 @@
 #define MAX(a,b) (((a)>(b))?(a):(b))
 #endif
 
-#define PP_OUTPUT_NB 3
+#define MODEL_OUTPUT_NB 3
 
 static od_pp_outBuffer_t out_detections[AI_OD_ST_YOLOX_PP_NB_ANCHORS *
                                         MAX(AI_OD_ST_YOLOX_PP_MAX_BOXES_LIMIT,
                                             AI_OD_ST_YOLOX_PP_L_GRID_WIDTH * AI_OD_ST_YOLOX_PP_L_GRID_HEIGHT +
                                             AI_OD_ST_YOLOX_PP_M_GRID_WIDTH * AI_OD_ST_YOLOX_PP_M_GRID_HEIGHT +
                                             AI_OD_ST_YOLOX_PP_S_GRID_WIDTH * AI_OD_ST_YOLOX_PP_S_GRID_HEIGHT)];
-/* will contain output index ordered by ascending output size */
-static size_t output_order_index[PP_OUTPUT_NB];
+/* will contain model output index ordered by ascending output size */
+static size_t output_order_index[MODEL_OUTPUT_NB];
 
 #include "sort.h"
 
@@ -43,7 +43,7 @@ int32_t app_postprocess_init(void *params_postprocess, stai_network_info *NN_Inf
   int32_t error = AI_OD_POSTPROCESS_ERROR_NO;
 
   assert(NN_Info);
-  sort_model_outputs(output_order_index, PP_OUTPUT_NB, NN_Info);
+  sort_model_outputs(output_order_index, MODEL_OUTPUT_NB, NN_Info);
 
   od_st_yolox_pp_static_param_t *params = (od_st_yolox_pp_static_param_t *) params_postprocess;
   params->nb_classes = AI_OD_ST_YOLOX_PP_NB_CLASSES;
@@ -67,7 +67,7 @@ int32_t app_postprocess_init(void *params_postprocess, stai_network_info *NN_Inf
 
 int32_t app_postprocess_run(void *pInput[], int nb_input, void *pOutput, void *pInput_param)
 {
-  assert(nb_input == PP_OUTPUT_NB);
+  assert(nb_input == MODEL_OUTPUT_NB);
   int32_t error = AI_OD_POSTPROCESS_ERROR_NO;
   ((od_st_yolox_pp_static_param_t *) pInput_param)->nb_detect = 0;
   od_pp_out_t *pObjDetOutput = (od_pp_out_t *) pOutput;