+++ Middlewares/stm32-vision-models-postprocessing/lib_vision_models_pp/Inc/od_ssd_pp_if.h
@@ -22,7 +22,6 @@
 typedef struct ssd_pp_in_centroid
 {
   void *pBoxes;
-  void *pAnchors;
   void *pScores;
 } od_ssd_pp_in_centroid_t;
 
@@ -33,18 +32,18 @@ typedef struct ssd_pp_static_param
 {
   int32_t   nb_classes;
   int32_t   nb_detections;
-  float32_t XY_scale;
-  float32_t WH_scale;
+  float32_t XY_inv_scale;
+  float32_t WH_inv_scale;
   int32_t   max_boxes_limit;
   float32_t conf_threshold;
   float32_t iou_threshold;
   int32_t   nb_detect;
-  void *scratchBuffer;
+  const void *pAnchors;
+  void *pScratchBuffer;
+  void *pScratchBufferSoftMax;
   float32_t boxe_scale;
-  float32_t anchor_scale;
   float32_t score_scale;
   int8_t boxe_zero_point;
-  int8_t anchor_zero_point;
   int8_t score_zero_point;
 } od_ssd_pp_static_param_t;
 
@@ -70,6 +69,19 @@ int32_t od_ssd_pp_reset(od_ssd_pp_static_param_t *pInput_static_param);
  * @retval Error code
  */
 int32_t od_ssd_pp_process(od_ssd_pp_in_centroid_t *pInput,
+                          od_pp_out_t *pOutput,
+                          od_ssd_pp_static_param_t *pInput_static_param);
+
+/*!
+ * @brief Object detector post processing : includes output detector remapping,
+ *        nms and score filtering for SSD int8 ouptputs from model
+ *
+ * @param [IN] Pointer on input structure pointing data
+ *             Pointer on output structure pointing data
+ *             pointer on static parameters
+ * @retval Error code
+ */
+int32_t od_ssd_pp_process_int8(od_ssd_pp_in_centroid_t *pInput,
                           od_pp_out_t *pOutput,
                           od_ssd_pp_static_param_t *pInput_static_param);